@@ -1,124 +1,136 @@
 {
-	"name": "@million23/next",
-	"version": "1.0.21",
-	"description": "Nextjs boilerplate with Tailwindcss, ShadcnUI, Framer Motion, Lucide, and Next Themes",
-	"author": "Gerald Chavez",
-	"bin": "./bin/cli.js",
-	"scripts": {
-		"dev": "set NODE_OPTIONS='--max-old-space-size=8192' && next dev",
-		"build": "next build",
-		"start": "next start --port 4321",
-		"lint": "next lint"
-	},
-	"keywords": [
-		"nexjs",
-		"radix",
-		"shadcn",
-		"boilerplate",
-		"framer-motion"
-	],
-	"dependencies": {
-		"@ai-sdk/mistral": "^0.0.8",
-		"@ai-sdk/openai": "^0.0.9",
-		"@auth/core": "0.31.0",
-		"@auth/xata-adapter": "^1.1.0",
-		"@clerk/nextjs": "^5.0.6",
-		"@clerk/themes": "^1.7.10",
-		"@hookform/resolvers": "^3.3.4",
-		"@huggingface/inference": "^2.6.6",
-		"@langchain/openai": "^0.0.14",
-		"@langchain/pinecone": "^0.0.3",
-		"@million23/next": "1.0.22",
-		"@pinecone-database/doc-splitter": "^0.0.1",
-		"@pinecone-database/pinecone": "^2.2.0",
-		"@radix-ui/react-accordion": "^1.1.2",
-		"@radix-ui/react-alert-dialog": "^1.0.5",
-		"@radix-ui/react-aspect-ratio": "^1.0.3",
-		"@radix-ui/react-avatar": "^1.0.4",
-		"@radix-ui/react-checkbox": "^1.0.4",
-		"@radix-ui/react-collapsible": "^1.0.3",
-		"@radix-ui/react-context-menu": "^2.1.5",
-		"@radix-ui/react-dialog": "^1.0.5",
-		"@radix-ui/react-dropdown-menu": "^2.0.6",
-		"@radix-ui/react-hover-card": "^1.0.7",
-		"@radix-ui/react-label": "^2.0.2",
-		"@radix-ui/react-menubar": "^1.0.4",
-		"@radix-ui/react-navigation-menu": "^1.1.4",
-		"@radix-ui/react-popover": "^1.0.7",
-		"@radix-ui/react-progress": "^1.0.3",
-		"@radix-ui/react-radio-group": "^1.1.3",
-		"@radix-ui/react-scroll-area": "^1.0.5",
-		"@radix-ui/react-select": "^2.0.0",
-		"@radix-ui/react-separator": "^1.0.3",
-		"@radix-ui/react-slider": "^1.1.2",
-		"@radix-ui/react-slot": "^1.0.2",
-		"@radix-ui/react-switch": "^1.0.3",
-		"@radix-ui/react-tabs": "^1.0.4",
-		"@radix-ui/react-toast": "^1.1.5",
-		"@radix-ui/react-toggle": "^1.0.3",
-		"@radix-ui/react-toggle-group": "^1.0.4",
-		"@radix-ui/react-tooltip": "^1.0.7",
-		"@tanstack/query-core": "^5.28.9",
-		"@tanstack/react-query": "^5.28.9",
-		"@uploadthing/react": "^6.5.1",
-		"@wojtekmaj/react-hooks": "^1.19.0",
-		"@xata.io/client": "^0.29.4",
-		"ai": "^3.1.8",
-		"axios": "^1.6.8",
-		"bcryptjs": "^2.4.3",
-		"bootstrap": "^5.3.3",
-		"class-variance-authority": "^0.7.0",
-		"clsx": "^2.1.0",
-		"cmdk": "^0.2.1",
-		"css-loader": "^6.10.0",
-		"docx-pdf": "^0.0.1",
-		"embla-carousel-react": "8.0.0-rc18",
-		"framer-motion": "^10.18.0",
-		"langchain": "^0.1.37",
-		"lucide-react": "^0.306.0",
-		"mammoth": "^1.7.2",
-		"md5": "^2.3.0",
-		"mini-css-extract-plugin": "^2.8.1",
-		"mkdirp": "^3.0.1",
-		"next": "14.2.3",
-		"next-auth": "5.0.0-beta.18",
-		"next-themes": "^0.2.1",
-		"openai": "^4.47.1",
-		"openai-edge": "^1.2.2",
-		"pdf-parse": "^1.1.1",
-		"react": "^18.3.1",
-		"react-day-picker": "^8.10.0",
-		"react-dom": "^18.3.1",
-		"react-hook-form": "^7.51.2",
-		"react-intersection-observer": "^9.8.1",
-		"react-loading-skeleton": "^3.4.0",
-		"react-resizable-panels": "^1.0.10",
-		"sonner": "^1.4.41",
-		"tailwind-merge": "^2.2.2",
-		"tailwindcss-animate": "^1.0.7",
-		"uploadthing": "^6.10.1",
-		"use-debounce": "^10.0.0",
-		"vaul": "^0.8.9",
-		"webpack": "^5.91.0",
-		"webpack-cli": "^5.1.4",
-		"zod": "^3.23.8"
-	},
-	"devDependencies": {
-		"@tanstack/eslint-plugin-query": "^5.28.6",
-		"@tanstack/react-query-devtools": "^5.28.10",
-		"@types/bcrypt": "^5.0.2",
-		"@types/bcryptjs": "^2.4.6",
-		"@types/md5": "^2.3.5",
-		"@types/node": "^20.11.30",
-		"@types/pdf-thumbnail": "^1.0.4",
-		"@types/react": "^18.2.73",
-		"@types/react-dom": "^18.2.23",
-		"autoprefixer": "^10.4.19",
-		"docx-to-pdf-axios": "^1.1.4",
-		"eslint": "^8.57.0",
-		"eslint-config-next": "14.2.3",
-		"postcss": "^8.4.38",
-		"tailwindcss": "^3.4.3",
-		"typescript": "^5.4.5"
-	}
+  "name": "rethink",
+  "version": "1.0.21",
+  "description": "Nextjs boilerplate with Tailwindcss, ShadcnUI, Framer Motion, Lucide, and Next Themes",
+  "author": "Rusty Ramos",
+  "bin": "./bin/cli.js",
+  "scripts": {
+    "dev": "next dev --turbopack",
+    "build": "next build",
+    "start": "next start --port 4321",
+    "lint": "next lint"
+  },
+  "keywords": [
+    "nexjs",
+    "radix",
+    "shadcn",
+    "boilerplate",
+    "framer-motion"
+  ],
+  "dependencies": {
+    "@ai-sdk/mistral": "^0.0.8",
+    "@ai-sdk/openai": "^1.3.21",
+    "@ai-sdk/react": "^1.2.11",
+    "@auth/core": "0.31.0",
+    "@auth/prisma-adapter": "^2.9.0",
+    "@auth/xata-adapter": "^1.9.0",
+    "@clerk/nextjs": "^5.7.5",
+    "@clerk/themes": "^1.7.20",
+    "@hookform/resolvers": "^3.10.0",
+    "@huggingface/inference": "^2.8.1",
+    "@langchain/openai": "^0.0.14",
+    "@langchain/pinecone": "^0.0.3",
+    "@pinecone-database/doc-splitter": "^0.0.1",
+    "@pinecone-database/pinecone": "^5.1.2",
+    "@prisma/client": "^6.7.0",
+    "@prisma/extension-accelerate": "^1.3.0",
+    "@radix-ui/react-accordion": "^1.2.8",
+    "@radix-ui/react-alert-dialog": "^1.1.11",
+    "@radix-ui/react-aspect-ratio": "^1.1.4",
+    "@radix-ui/react-avatar": "^1.1.7",
+    "@radix-ui/react-checkbox": "^1.2.3",
+    "@radix-ui/react-collapsible": "^1.1.8",
+    "@radix-ui/react-context-menu": "^2.2.12",
+    "@radix-ui/react-dialog": "^1.1.11",
+    "@radix-ui/react-dropdown-menu": "^2.1.12",
+    "@radix-ui/react-hover-card": "^1.1.11",
+    "@radix-ui/react-label": "^2.1.4",
+    "@radix-ui/react-menubar": "^1.1.12",
+    "@radix-ui/react-navigation-menu": "^1.2.10",
+    "@radix-ui/react-popover": "^1.1.11",
+    "@radix-ui/react-progress": "^1.1.4",
+    "@radix-ui/react-radio-group": "^1.3.4",
+    "@radix-ui/react-scroll-area": "^1.2.6",
+    "@radix-ui/react-select": "^2.2.2",
+    "@radix-ui/react-separator": "^1.1.4",
+    "@radix-ui/react-slider": "^1.3.2",
+    "@radix-ui/react-slot": "^1.2.0",
+    "@radix-ui/react-switch": "^1.2.2",
+    "@radix-ui/react-tabs": "^1.1.9",
+    "@radix-ui/react-toast": "^1.2.11",
+    "@radix-ui/react-toggle": "^1.1.6",
+    "@radix-ui/react-toggle-group": "^1.1.7",
+    "@radix-ui/react-tooltip": "^1.2.4",
+    "@tailwindcss/postcss": "^4.1.5",
+    "@tanstack/query-core": "^5.74.9",
+    "@tanstack/react-query": "^5.74.11",
+    "@uploadthing/react": "^6.8.0",
+    "@wojtekmaj/react-hooks": "^1.22.0",
+    "@xata.io/client": "^0.0.0-next.vfbf85e03c9dbded87f0e347cc3850e811f46ef97",
+    "ai": "^4.3.13",
+    "axios": "^1.9.0",
+    "bcryptjs": "^2.4.3",
+    "bootstrap": "^5.3.5",
+    "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
+    "cmdk": "^0.2.1",
+    "css-loader": "^6.11.0",
+    "docx-pdf": "^0.0.1",
+    "dotenv": "^16.5.0",
+    "drizzle-orm": "^0.43.1",
+    "embla-carousel-react": "8.0.0-rc18",
+    "framer-motion": "^10.18.0",
+    "langchain": "^0.1.37",
+    "lucide-react": "^0.306.0",
+    "mammoth": "^1.9.0",
+    "md5": "^2.3.0",
+    "mini-css-extract-plugin": "^2.9.2",
+    "mkdirp": "^3.0.1",
+    "next": "15.3.1",
+    "next-auth": "5.0.0-beta.25",
+    "next-themes": "^0.2.1",
+    "openai": "^4.96.2",
+    "openai-edge": "^1.2.3",
+    "pdf-parse": "^1.1.1",
+    "prisma": "^6.7.0",
+    "react": "19.1.0",
+    "react-day-picker": "^8.10.1",
+    "react-dom": "19.1.0",
+    "react-hook-form": "^7.56.1",
+    "react-intersection-observer": "^9.16.0",
+    "react-loading-skeleton": "^3.5.0",
+    "react-resizable-panels": "^1.0.10",
+    "sonner": "^1.7.4",
+    "tailwind-merge": "^2.6.0",
+    "tailwindcss-animate": "^1.0.7",
+    "uploadthing": "7",
+    "use-debounce": "^10.0.4",
+    "vaul": "^0.8.9",
+    "webpack": "^5.99.7",
+    "webpack-cli": "^5.1.4",
+    "zod": "^3.24.3"
+  },
+  "devDependencies": {
+    "@tanstack/eslint-plugin-query": "^5.74.7",
+    "@tanstack/react-query-devtools": "^5.74.11",
+    "@types/bcrypt": "^5.0.2",
+    "@types/bcryptjs": "^2.4.6",
+    "@types/md5": "^2.3.5",
+    "@types/node": "^20.17.32",
+    "@types/pdf-thumbnail": "^1.0.4",
+    "@types/react": "19.1.2",
+    "@types/react-dom": "19.1.2",
+    "autoprefixer": "^10.4.21",
+    "docx-to-pdf-axios": "^1.1.4",
+    "drizzle-kit": "^0.31.0",
+    "eslint": "^8.57.1",
+    "eslint-config-next": "15.3.1",
+    "postcss": "^8.5.3",
+    "tailwindcss": "4.1.5",
+    "typescript": "^5.8.3"
+  },
+  "overrides": {
+    "@types/react": "19.1.2",
+    "@types/react-dom": "19.1.2"
+  }
 }
