@@ -1,4 +1,4 @@
-import { handlers } from "@/auth" // Referring to the auth.ts we just created
+import { handlers } from "../../../../auth" // Referring to the auth.ts we just created
 
 
 export const { GET, POST } = handlers
