@@ -3,10 +3,8 @@ import { createRouteHandler } from "uploadthing/next";
 import { ourFileRouter } from "./core";
 
 // Export routes for Next App Router
-export const { GET, POST } = createRouteHandler({
+const handler = createRouteHandler({
   router: ourFileRouter,
-  config: {
-     uploadthingSecret: process.env.UPLOADTHING_SECRET,
-     uploadthingId: process.env.UPLOADTHING_APP_ID,
-  },
 });
+
+export { handler as GET, handler as POST };
