@@ -1,7 +1,6 @@
-import { withUt } from 'uploadthing/tw';
 import type { Config } from 'tailwindcss';
 
-const config: Config = withUt({
+const config: Config = {
   darkMode: ['class'],
   content: [
     './pages/**/*.{ts,tsx}',
@@ -79,6 +78,6 @@ const config: Config = withUt({
     },
   },
   plugins: [require('tailwindcss-animate')],
-});
+};
 
 export default config;
